pooled pg clients and basic user page with search OR create new functionality.
